+++ tests/Converter/Time/PhpTimeConverterTest.php
@@ -29,16 +29,14 @@ class PhpTimeConverterTest extends TestCase
         $maskMid = BigInteger::fromBase('ffff', 16);
         $maskHi = BigInteger::fromBase('0fff', 16);
 
-        $expectedArray = [
-            'low' => sprintf('%08s', $calculatedTime->and($maskLow)->toBase(16)),
-            'mid' => sprintf('%04s', $calculatedTime->shiftedRight(32)->and($maskMid)->toBase(16)),
-            'hi' => sprintf('%04s', $calculatedTime->shiftedRight(48)->and($maskHi)->toBase(16)),
-        ];
+        $expected = sprintf('%04s', $calculatedTime->shiftedRight(48)->and($maskHi)->toBase(16));
+        $expected .= sprintf('%04s', $calculatedTime->shiftedRight(32)->and($maskMid)->toBase(16));
+        $expected .= sprintf('%08s', $calculatedTime->and($maskLow)->toBase(16));
 
         $converter = new PhpTimeConverter();
         $returned = $converter->calculateTime((string) $seconds, (string) $microSeconds);
 
-        $this->assertSame($expectedArray, $returned);
+        $this->assertSame($expected, $returned->toString());
     }
 
     public function testCalculateTimeThrowsExceptionWhenSecondsIsNotOnlyDigits(): void
@@ -134,11 +132,9 @@ class PhpTimeConverterTest extends TestCase
     }
 
     /**
-     * @param string[] $expected
-     *
      * @dataProvider provideCalculateTime
      */
-    public function testCalculateTime(string $seconds, string $microSeconds, array $expected): void
+    public function testCalculateTime(string $seconds, string $microSeconds, string $expected): void
     {
         $calculator = new BrickMathCalculator();
         $fallbackConverter = new GenericTimeConverter($calculator);
@@ -146,7 +142,7 @@ class PhpTimeConverterTest extends TestCase
 
         $result = $converter->calculateTime($seconds, $microSeconds);
 
-        $this->assertSame($expected, $result);
+        $this->assertSame($expected, $result->toString());
     }
 
     /**
@@ -158,29 +154,17 @@ class PhpTimeConverterTest extends TestCase
             [
                 'seconds' => '-12219146756',
                 'microSeconds' => '0',
-                'expected' => [
-                    'low' => '0901e600',
-                    'mid' => '0154',
-                    'hi' => '0000',
-                ],
+                'expected' => '000001540901e600',
             ],
             [
                 'seconds' => '103072857659',
                 'microseconds' => '999999',
-                'expected' => [
-                    'low' => 'ff9785f6',
-                    'mid' => 'ffff',
-                    'hi' => '0fff',
-                ],
+                'expected' => '0fffffffff9785f6',
             ],
             [
                 'seconds' => '1578612359',
                 'microseconds' => '521023',
-                'expected' => [
-                    'low' => '64c71df6',
-                    'mid' => '3337',
-                    'hi' => '01ea',
-                ],
+                'expected' => '01ea333764c71df6',
             ],
 
             // This is the earliest possible date supported by v1 UUIDs:
@@ -188,11 +172,7 @@ class PhpTimeConverterTest extends TestCase
             [
                 'seconds' => '-12219292800',
                 'microSeconds' => '0',
-                'expected' => [
-                    'low' => '00000000',
-                    'mid' => '0000',
-                    'hi' => '0000',
-                ],
+                'expected' => '0000000000000000',
             ],
 
             // This is the last possible time supported by v1 UUIDs:
@@ -200,11 +180,7 @@ class PhpTimeConverterTest extends TestCase
             [
                 'seconds' => '1832455114570',
                 'microseconds' => '955161',
-                'expected' => [
-                    'low' => 'fffffffa',
-                    'mid' => 'ffff',
-                    'hi' => 'ffff',
-                ],
+                'expected' => 'fffffffffffffffa',
             ],
         ];
     }